Rewrite notification API with unified endpoint
