+++ RemoteControlApi/wwwroot/receive.html
@@ -103,7 +103,7 @@
         }
 
         async function load() {
-            const res = await fetch(urlFrom('api/control/get-notifications'));
+            const res = await fetch(urlFrom('api/notifications'));
             if (!res.ok) {
                 list.innerHTML = `<li>Load failed: ${res.status}</li>`;
                 return;
@@ -116,7 +116,7 @@
         load();
 
         // SSE phải dùng URL đúng base
-        const evt = new EventSource(urlFrom('api/control/notifications-stream'));
+        const evt = new EventSource(urlFrom('api/notifications/stream'));
         evt.onmessage = e => add(JSON.parse(e.data));
         evt.onerror = () => {
             // optional: auto-reconnect (EventSource tự reconnect mặc định)
