+++ RemoteControlApi/wwwroot/send.html
@@ -77,7 +77,7 @@
 
             try {
                 const data = new FormData(form);
-                const res = await fetch(urlFrom('api/control/send-notification'), {
+                const res = await fetch(urlFrom('api/notifications'), {
                     method: 'POST',
                     body: data
                 });
